Discard data from input buffers after errors
In case of protocol errors, the connection is no longer usable.  If
there's data remaining in buffers, client.available() can still return
non-zero, even if the underlying connection is already closed.  Still,
because of the error, we won't know how to interpret the remaining data,
so this commit makes the library read and discard it.  This is required
to not confuse any other logic of the library.